Update v0.1.10
- Support y4m writing for recon file in the encoder.
- Bug fix on meta data writing.
- Add full range flag.
- Add a few assert for tile, metadata and etc.
- Update for coding the systax elements with _minus1.
- Fix the range of DC differenece from 16bits to 17bits.
- Add NEON func for SAD and DIFF.
- Add AVX func for SAD computations.

+++ src/oapv_vlc.h
@@ -62,6 +62,6 @@ int  oapvd_vlc_tile_header(oapv_bs_t* bs, oapvd_ctx_t* ctx, oapv_th_t* th);
 int  oapvd_vlc_tile_dummy_data(oapv_bs_t* bs);
 int  oapvd_vlc_metadata(oapv_bs_t* bs, u32 pbu_size, oapvm_t mid, int group_id);
 int  oapvd_vlc_filler(oapv_bs_t* bs, u32 filler_size);
-int  oapvd_vlc_dc_coeff(oapvd_ctx_t* ctx, oapvd_core_t* core, oapv_bs_t* bs, s16* dc_diff, int c);
+int  oapvd_vlc_dc_coeff(oapvd_ctx_t* ctx, oapvd_core_t* core, oapv_bs_t* bs, int* dc_diff, int c);
 int  oapvd_vlc_ac_coeff(oapvd_ctx_t* ctx, oapvd_core_t* core, oapv_bs_t* bs, s16* coef, int c);
 #endif /* _OAPV_VLC_H_ */
